Move VersionType to interfaces module, since that's the only place it's used, other than to pass it in as arg.
