+++ src/ogd/common/storage/interfaces/BigQueryInterface.py
@@ -19,7 +19,7 @@ from ogd.common.configs.storage.BigQueryConfig import BigQueryConfig
 from ogd.common.models.SemanticVersion import SemanticVersion
 from ogd.common.models.enums.IDMode import IDMode
 from ogd.common.filters.FilterMode import FilterMode
-from ogd.common.models.enums.VersionType import VersionType
+from ogd.common.storage.interfaces.VersionType import VersionType
 from ogd.common.storage.interfaces.Interface import Interface
 from ogd.common.storage.connectors.BigQueryConnector import BigQueryConnector
 from ogd.common.utils.Logger import Logger
