+++ src/ogd/common/storage/interfaces/CSVInterface.py
@@ -13,7 +13,7 @@ from ogd.common.configs.storage.FileStoreConfig import FileStoreConfig
 from ogd.common.models.enums.ExportMode import ExportMode
 from ogd.common.models.enums.IDMode import IDMode
 from ogd.common.filters.FilterMode import FilterMode
-from ogd.common.models.enums.VersionType import VersionType
+from ogd.common.storage.interfaces.VersionType import VersionType
 from ogd.common.models.SemanticVersion import SemanticVersion
 from ogd.common.storage.interfaces.Interface import Interface
 from ogd.common.storage.connectors.CSVConnector import CSVConnector
