+++ src/ogd/common/storage/interfaces/Interface.py
@@ -20,7 +20,7 @@ from ogd.common.models.Feature import Feature
 from ogd.common.models.FeatureSet import FeatureSet
 from ogd.common.models.enums.IDMode import IDMode
 from ogd.common.filters.FilterMode import FilterMode
-from ogd.common.models.enums.VersionType import VersionType
+from ogd.common.storage.interfaces.VersionType import VersionType
 from ogd.common.models.SemanticVersion import SemanticVersion
 from ogd.common.configs.DataTableConfig import DataTableConfig
 from ogd.common.schemas.tables.EventTableSchema import EventTableSchema
