+++ tests/cases/storage/interfaces/BigQueryInterfaceSuite/OGDBigQueryCase.py
@@ -13,7 +13,7 @@ from ogd.common.configs.DataTableConfig import DataTableConfig
 from ogd.common.configs.TestConfig import TestConfig
 from ogd.common.models.enums.IDMode import IDMode
 from ogd.common.filters.FilterMode import FilterMode
-from ogd.common.models.enums.VersionType import VersionType
+from ogd.common.storage.interfaces.VersionType import VersionType
 from ogd.common.schemas.tables.EventTableSchema import EventTableSchema
 from ogd.common.storage.interfaces.BigQueryInterface import BigQueryInterface, ParamaterizedClause
 from ogd.common.utils.Logger import Logger
